Shapes v2 review fixes: Fragment label, seed-rule tests, type/wording nits
1. Formation label gap. A zero-origin token (legal per §4 — e.g. the 0-origin
   50 ETH half of a split Direct 100) read as "Composed", stating a history it
   never had. Add a fourth label: originCount == 0 => "Fragment", in the Solidity
   renderer, the TS canonical renderer, the frontend badge, and spec §2. Covered
   by a new Fragment fixture and an on-chain provenance assertion (split a Direct
   100 into two 50s; the second half is a Fragment).

2. Deterministic seed rule now tested end to end. Exact derivation was already
   asserted on-chain; add (a) invariance across vm.roll/warp/prevrandao/fee/chainId
   (proves no block data leaks into a child seed, so a split cannot be re-rolled
   per block), and (b) Solidity<->TypeScript parity: a shared derivation helper
   preview/src/decomposeSeed.ts (keccak256(abi.encodePacked(parentSeed, i)), the
   same rule the frontend previews with), fixtures generated from it, and a Foundry
   test recomputing the Solidity derivation against them.

3. Nits. Composed event originCount aligned to uint32 (matches Decomposed's
   uint32[]); spec §9.2 states the survivor is included in the "none Black" check;
   spec §16 phase 3 notes the child-seed derivation and its TS twin land there.

140 tests, 0 failed (4 fork-only skipped). Preview typechecks.

+++ src/Shapes.sol
@@ -372,7 +372,7 @@ contract Shapes is ERC721, ReentrancyGuard, Ownable, IShapes, IERC4906 {
         s.denomIndex = uint8(newIndex);
         s.originCount = uint32(origins); // <= total/UNIT <= 10000 by the capacity invariant
 
-        emit Composed(survivorId, burnIds, uint8(newIndex), origins);
+        emit Composed(survivorId, burnIds, uint8(newIndex), uint32(origins));
         emit MetadataUpdate(survivorId);
         return survivorId;
     }
